+++ tests/fields/test_radio.py
@@ -1,3 +1,6 @@
+from tests.common import DummyPostData
+
+from wtforms import validators
 from wtforms.fields import RadioField
 from wtforms.form import Form
 
@@ -9,6 +12,10 @@ def make_form(name="F", **fields):
 class F(Form):
     a = RadioField(choices=[("a", "hello"), ("b", "bye")], default="a")
     b = RadioField(choices=[(1, "Item 1"), (2, "Item 2")], coerce=int)
+    c = RadioField(
+        choices=[("a", "Item 1"), ("b", "Item 2")],
+        validators=[validators.InputRequired()],
+    )
 
 
 def test_radio_field():
@@ -73,4 +80,41 @@ def test_callable_choices():
         '<li><input id="a-1" name="a" type="radio" value="b"> '
         '<label for="a-1">bye</label></li>'
         "</ul>"
+    )
+
+
+def test_required_flag():
+    form = F()
+    assert form.c() == (
+        '<ul id="c">'
+        '<li><input id="c-0" name="c" required type="radio" value="a"> '
+        '<label for="c-0">Item 1</label></li>'
+        '<li><input id="c-1" name="c" required type="radio" value="b"> '
+        '<label for="c-1">Item 2</label></li>'
+        "</ul>"
+    )
+
+
+def test_required_validator():
+    form = F(DummyPostData(b=1, c="a"))
+    assert form.validate()
+    assert form.c.errors == []
+    form = F(DummyPostData(b=1))
+    assert form.validate() is False
+    assert form.c.errors == ["This field is required."]
+
+
+def test_render_kw_preserved():
+    F = make_form(
+        a=RadioField(
+            choices=[(True, "yes"), (False, "no")], render_kw=dict(disabled=True)
+        )
+    )
+    form = F()
+    assert form.a() == (
+        '<ul disabled id="a">'
+        '<li><input disabled id="a-0" name="a" type="radio" value="True"> '
+        '<label for="a-0">yes</label></li><li>'
+        '<input disabled id="a-1" name="a" type="radio" value="False"> '
+        '<label for="a-1">no</label></li></ul>'
     )